Added Security Config to other classes

+++ src/main/java/org/example/finalprojecttuwaiq/Service/DraftRequirementService.java
@@ -2,8 +2,12 @@ package org.example.finalprojecttuwaiq.Service;
 
 import lombok.RequiredArgsConstructor;
 import org.example.finalprojecttuwaiq.Api.ApiException;
+import org.example.finalprojecttuwaiq.Model.BA;
 import org.example.finalprojecttuwaiq.Model.DraftRequirement;
+import org.example.finalprojecttuwaiq.Model.Project;
+import org.example.finalprojecttuwaiq.Repository.BARepository;
 import org.example.finalprojecttuwaiq.Repository.DraftRequirementRepository;
+import org.example.finalprojecttuwaiq.Repository.ProjectRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -12,7 +16,8 @@ import java.util.List;
 @RequiredArgsConstructor
 public class DraftRequirementService {
     private final DraftRequirementRepository draftRequirementRepository;
-
+    private final BARepository baRepository;
+    private final ProjectRepository projectRepository;
     public  List<DraftRequirement> getDraftRequirementByProjectId(Integer project_id){
         return draftRequirementRepository.findDraftRequirementByProject_id(project_id);
     }
@@ -20,15 +25,39 @@ public class DraftRequirementService {
         return draftRequirementRepository.findAll();
     }
 
-    public  DraftRequirement getDraftById(Integer draft_id){
-        return draftRequirementRepository.findDraftRequirementById(draft_id);
-    }
-
-    //TODO: Reject Draft And Delete it from DB By Id
-    public void rejectDraft(Integer draft_id){
+    public  DraftRequirement getDraftById(Integer ba_id,Integer draft_id){
         DraftRequirement draftRequirement = draftRequirementRepository.findDraftRequirementById(draft_id);
         if (draftRequirement == null)
             throw new ApiException("Draft Not Found");
+        Project project = projectRepository.findProjectById(draftRequirement.getProject_id());
+        if (project == null)
+            throw new ApiException("project not found");
+        BA ba = baRepository.findBAById(ba_id);
+        if (ba == null)
+            throw new ApiException("BA not found");
+
+        if (!project.getBas().contains(ba))
+            throw new ApiException("Not Authorized");
+        return draftRequirement;
+    }
+
+    //TODO: Reject Draft And Delete it from DB By Id
+    public void rejectDraft(Integer ba_id,Integer draft_id){
+        DraftRequirement draftRequirement = draftRequirementRepository.findDraftRequirementById(draft_id);
+        if (draftRequirement == null)
+            throw new ApiException("Draft Not Found");
+        Project project = projectRepository.findProjectById(draftRequirement.getProject_id());
+        if (project == null)
+            throw new ApiException("project not found");
+        BA ba = baRepository.findBAById(ba_id);
+        if (ba == null)
+            throw new ApiException("BA not found");
+
+        if (!project.getBas().contains(ba))
+            throw new ApiException("Not Authorized");
+
+
+
         draftRequirementRepository.delete(draftRequirement);
     }
 }
